backend-v1/requirements.txt
# FastAPI Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
motor>=3.3.2
pymongo>=4.6.1

# Vector Database
qdrant-client>=1.7.0

# Embeddings & ML
sentence-transformers>=2.3.1
numpy>=1.24.0

# LLM (google-genai is the new unified SDK)
google-genai>=1.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
pytesseract>=0.3.10
Pillow>=10.0.0
pypdfium2>=4.30.0

# File Storage (Cloudflare R2 / S3-compatible)
boto3>=1.34.0

# Utilities
pytz>=2023.3
